+++ src/test/unit/init/init_test.ts
@@ -25,6 +25,8 @@ const temp = tempMod.track();
 
 
 const isPlatformWin = /^win/.test(process.platform);
+const uname = childProcess.execSync('uname -s').toString();
+const isMinGw = !!/^mingw/i.test(uname);
 
 function stripAnsi(str: string) {
   const ansiRegex =
@@ -180,7 +182,7 @@ suite('init', () => {
 
     test('works with the default yeoman environment', async () => {
       sandbox.stub(inquirer, 'prompt').returns(Promise.resolve({
-        generatorName: 'TEST',
+        foo: 'TEST',
       }));
       (polymerInit as any).runGenerator = function() {};
       const error = await invertPromise(polymerInit.promptGeneratorSelection());
@@ -191,7 +193,7 @@ suite('init', () => {
         'prompts with a list to get generatorName property from user';
     test(testName, async () => {
       const promptStub = sandbox.stub(inquirer, 'prompt')
-                             .returns(Promise.resolve({generatorName: 'TEST'}));
+                             .returns(Promise.resolve({foo: 'TEST'}));
       try {
         await polymerInit.promptGeneratorSelection({env: yeomanEnvMock});
       } catch (error) {
@@ -199,7 +201,6 @@ suite('init', () => {
       }
       assert.isTrue(promptStub.calledOnce);
       assert.equal(promptStub.firstCall.args[0][0].type, 'list');
-      assert.equal(promptStub.firstCall.args[0][0].name, 'generatorName');
       assert.equal(
           promptStub.firstCall.args[0][0].message,
           'Which starter template would you like to use?');
@@ -207,7 +208,7 @@ suite('init', () => {
 
     test('prompts with a list of all registered generators', async () => {
       const promptStub = sandbox.stub(inquirer, 'prompt')
-                             .returns(Promise.resolve({generatorName: 'TEST'}));
+                             .returns(Promise.resolve({foo: 'TEST'}));
       try {
         await polymerInit.promptGeneratorSelection({env: yeomanEnvMock});
       } catch (error) {
@@ -235,7 +236,7 @@ suite('init', () => {
       const yeomanEnv = new YeomanEnvironment();
       const promptStub =
           sandbox.stub(inquirer, 'prompt').returns(Promise.resolve({
-            generatorName: 'TEST',
+            foo: 'TEST',
           }));
       helpers.registerDependencies(yeomanEnv, [[
                                      helpers.createDummyGenerator(),
@@ -257,7 +258,7 @@ suite('init', () => {
 
     test('prompts the user with a list', async () => {
       const promptStub = sandbox.stub(inquirer, 'prompt')
-                             .returns(Promise.resolve({generatorName: 'TEST'}));
+                             .returns(Promise.resolve({foo: 'TEST'}));
 
       try {
         await polymerInit.promptGeneratorSelection({env: yeomanEnvMock});
@@ -268,12 +269,10 @@ suite('init', () => {
       assert.equal(promptStub.firstCall.args[0][0].type, 'list');
     });
 
-    if (isPlatformWin) {
+    if (isPlatformWin && isMinGw) {
       test('prompts with a rawlist if being used in MinGW shell', async () => {
-        const promptStub =
-            sandbox.stub(inquirer, 'prompt').returns(Promise.resolve({
-              generatorName: 'TEST'
-            }));
+        const promptStub = sandbox.stub(inquirer, 'prompt')
+                               .returns(Promise.resolve({foo: 'TEST'}));
         sandbox.stub(childProcess, 'execSync')
             .withArgs('uname -s')
             .returns('mingw');