+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 - Added `--fix` option to `polymer lint`. When passed, some warnings with simple mechanical solutions will be fixed.
+  - Also supports warnings which can be addressed with less-safe changes via
+    an interactive prompt and the `--edits` flag. See `polymer lint --help` for
+    more info.
 - `build` Added a CLI argument for setting the `basePath` option: `--base-path`.
 <!-- Add new, unreleased items here. -->